Promote BioAssay.sequenceReadCount to a long

+++ gemma-core/src/main/java/ubic/gemma/core/analysis/expression/diff/LinearModelAnalyzer.java
@@ -537,7 +537,7 @@ public class LinearModelAnalyzer extends AbstractDifferentialExpressionAnalyzer 
                 assert bas.size() == 1;
                 BioAssay ba = bas.iterator().next();
 
-                Integer sequenceReadCount = ba.getSequenceReadCount();
+                Long sequenceReadCount = ba.getSequenceReadCount();
                 if ( !ba.getIsOutlier() && ( sequenceReadCount == null || sequenceReadCount == 0 ) ) {
                     // double check.
                     Double[] col = dmatrix.getColumn( i );
@@ -549,7 +549,7 @@ public class LinearModelAnalyzer extends AbstractDifferentialExpressionAnalyzer 
                     }
                 }
 
-                if ( sequenceReadCount == null ) sequenceReadCount = 0;
+                if ( sequenceReadCount == null ) sequenceReadCount = 0L;
 
                 librarySize.set( i, sequenceReadCount );
             }
